+++ docs/source/changelog.rst
@@ -8,6 +8,7 @@ ChangeLog
 HEAD
 ****
 
+* Fix value of the ``authorityKeyIdentifier`` extension when signing certificates with an intermediate CA.
 * Validate DNSNames when parsing general names. This means that signing a certificate with CommonName 
   that is not a valid domain name fails if it should also be added as subjectAltName (see ``--cn-in-san``
   option).